transaction class added
transaction class added with green spec file

the account class deposit and withdrawal methods have been amended to 
pass the previous transactions balance to the transasction being 
initialised. This allows the transasction object to set its own balance.

This logic means the balance is a staic attribute of the transaction and 
that it cannot be amended after initialisation.