refactored currency validation
The currency? validator behavious was working in some situatios but 
failing in others. This was due to float data types throwing spurious 
results. The validator has been updated to use an instance of BigDecimal 
before validating the amount which is converterd to a string.

This commit also includes better test coverage for the validation 
module.